pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "taro-core"
authors = [{name = "Stan Svec", email = "dev@stansvec.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
requires-python = ">=3.8"
dynamic = ["version", "description"]
dependencies = [
    "portalocker>=2.6.0",
    "pygrok>=1.0.0",
    "python-dateutil>=2.8.2",
    "PyYAML>=5.1.2",
    "urllib3>=1.26.2",
]

[project.optional-dependencies]
test = ["pytest"]

[tool.flit.module]
name = "tarotools.taro"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = "src"

[project.urls]
Home = "https://github.com/tarotools/taro-core"
